+++ benchmarks/verus/apps/files/apps_test_4160.rs
@@ -19,12 +19,12 @@ spec fn calculate_deposit(initial: int, years: int) -> int
 // </vc-helpers>
 
 // <vc-spec>
-fn solve(x: int) -> (years: int)
+fn solve(x: i8) -> (years: i8)
     requires x >= 101
     ensures 
         years >= 0 &&
-        calculate_deposit(100, years) >= x &&
-        (years == 0 || calculate_deposit(100, years - 1) < x)
+        calculate_deposit(100, years as int) >= x as int &&
+        (years == 0 || calculate_deposit(100, (years - 1) as int) < x as int)
 // </vc-spec>
 // <vc-code>
 {
